+++ src/ANAFEntity/InactiveInfo.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace EdituraEDU\ANAF\ANAFEntity;
 
 /**
@@ -6,24 +7,25 @@ namespace EdituraEDU\ANAF\ANAFEntity;
  */
 class InactiveInfo
 {
-    public string $dataInactivare="";
-    public string $dataReactivare="";
-    public string $dataPublicare="";
-    public string $dataRadiere="";
-    public bool $statusInactivi=false;
+    public string $dataInactivare = "";
+    public string $dataReactivare = "";
+    public string $dataPublicare = "";
+    public string $dataRadiere = "";
+    public bool $statusInactivi = false;
+
     /**
-     * Similar to @see Entity::CreateFromParsed
-     * @param \stdClass $parsedData
+     * Similar to @param \stdClass $parsedData
      * @return InactiveInfo
+     * @see Entity::CreateFromParsed
      */
-    public static function CreateFromParsed(\stdClass $parsedData):InactiveInfo
+    public static function CreateFromParsed(\stdClass $parsedData): InactiveInfo
     {
         $inactiveInfo = new InactiveInfo();
         $inactiveInfo->dataInactivare = $parsedData->dataInactivare ?? "";
-        $inactiveInfo->dataReactivare = $parsedData->dataReactivare??"";
-        $inactiveInfo->dataPublicare = $parsedData->dataPublicare??"";
-        $inactiveInfo->dataRadiere = $parsedData->dataRadiere??"";
-        $inactiveInfo->statusInactivi = $parsedData->statusInactivi??"";
+        $inactiveInfo->dataReactivare = $parsedData->dataReactivare ?? "";
+        $inactiveInfo->dataPublicare = $parsedData->dataPublicare ?? "";
+        $inactiveInfo->dataRadiere = $parsedData->dataRadiere ?? "";
+        $inactiveInfo->statusInactivi = $parsedData->statusInactivi ?? "";
         return $inactiveInfo;
     }
 }